+++ src/lib/core/common/services/events.service.ts
@@ -8,8 +8,8 @@ export class AtEvents {
     /**
      * Subscribe to an event topic. Events that get posted to that topic will trigger the provided handler.
      *
-     * @param {string} topic the topic to subscribe to
-     * @param {function} handler the event handler
+     * @param topic the topic to subscribe to
+     * @param handler the event handler
      */
     subscribe(topic: string, ...handlers: Function[]) {
         if (!this._channels[topic]) {
@@ -23,8 +23,8 @@ export class AtEvents {
     /**
      * Unsubscribe from the given topic. Your handler will no longer receive events published to this topic.
      *
-     * @param {string} topic the topic to unsubscribe from
-     * @param {function} handler the event handler
+     * @param topic the topic to unsubscribe from
+     * @param handler the event handler
      *
      * @return true if a handler was removed
      */
@@ -62,8 +62,8 @@ export class AtEvents {
     /**
      * Publish an event to the given topic.
      *
-     * @param {string} topic the topic to publish to
-     * @param {any} eventData the data to send as the event
+     * @param topic the topic to publish to
+     * @param eventData the data to send as the event
      */
     publish(topic: string, ...args: any[]) {
         var t = this._channels[topic];